refactor(host-detail): collapse NodeDetails into EntityDetailModal (#578)
There were two near-identical host-detail modals: the graph's NodeDetails
and the entity panel EntityDetailModal (monitor/conversation/analysis).
After #575 both embedded HostIdentitySection, but the shells still diverged
— so a host clicked in the graph looked "almost but not quite" like the same
host clicked in monitor, and geo/other bits differed by surface.

Unify on EntityDetailModal as the single shell. When given graph context
(`graphNode`/`graphEdges`), it now renders the graph-only pieces that used to
live in NodeDetails:
- GraphNodeDetailsSection: ghost-node warning, traffic counters, protocol
  chips, and the per-peer Connections table (from graph edges).
- GraphHistoryTab: the cross-capture role-trail history (its own tab).
- Service-role log tabs (DNS/web/…) and the change-event highlight banner.

Migrate all three NodeDetails call sites — NetworkDiagramPage, ComparePage,
SnapshotDetailModal (monitor) — to EntityDetailModal via a small
`graphNodeEntity(node)` helper (L2 → DEVICE/MAC, else IP). Delete NodeDetails
and its CSS.

Side fix: the .tp-nested-modal z-index rules lived in NodeDetails.css and
only loaded when the graph panel was mounted; they now ship with
EntityDetailModal.css, so the add-evidence/explainer/role modals stack
correctly on every surface, not just the graph.

Frontend-only. Type-check clean; all 145 tests pass. Playwright-verified the
graph node click opens the unified modal with Identity + evidence axes +
geolocation + Connections + Protocols + History/service tabs.
