+++ frontend/src/components/common/EntityDetailModal/index.ts
@@ -1,1 +1,2 @@
 export { EntityDetailModal } from './EntityDetailModal';
+export { graphNodeEntity } from './graphNodeEntity';
