+++ frontend/src/components/common/EntityDetailModal/types.ts
@@ -1,7 +1,10 @@
 import type { EntityType } from '@/features/notes/services/entityNotesService';
 import type { NetworkSnapshot } from '@/features/monitor/types/monitor.types';
+import type { GraphNode, GraphEdge } from '@/features/network/types';
+import type { NodeHighlight } from '@/components/network/NetworkGraph/NetworkGraph';
 
-export type Tab = 'details' | 'notes';
+/** Base tabs; graph context adds a 'history' tab and dynamic `svc:<role>` service tabs. */
+export type Tab = 'details' | 'notes' | 'history' | string;
 
 export interface HostClassification {
   ip: string | null;
@@ -56,4 +59,18 @@ export interface EntityDetailModalProps {
   snapshots?: NetworkSnapshot[];
   onClose: () => void;
   zIndex?: number;
+
+  // ── Graph context (network-diagram surfaces) ─────────────────────────────────
+  // When `graphNode` is supplied, the modal renders the network-graph host detail: measured traffic
+  // counters, protocol chips, a per-peer Connections table (from `graphEdges`), the History tab, and
+  // any service-role log tabs — the pieces that used to live in the standalone NodeDetails. Omit them
+  // and the modal is the plain entity panel (monitor drift / analysis).
+  /** The clicked graph node; presence switches on all graph-only sections. */
+  graphNode?: GraphNode;
+  /** All graph edges, for building this node's per-peer Connections table. */
+  graphEdges?: GraphEdge[];
+  /** Change-event highlight banner (Monitor snapshot / Compare diff context). */
+  changeHighlight?: NodeHighlight;
+  /** Navigate away (peer-row / history-row links). Given a path; the caller routes + closes. */
+  onNavigate?: (path: string) => void;
 }
