--- frontend/src/components/network/NodeDetails/NodeDetails.css
@@ -1,96 +1,0 @@
-.node-details-panel {
-  position: sticky;
-  top: 20px;
-  max-height: calc(100vh - 100px);
-  overflow-y: auto;
-}
-
-.stats-table {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.stats-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.25rem 0;
-}
-
-.stats-label {
-  font-size: 0.875rem;
-  color: var(--tp-text-muted, #6c757d);
-}
-
-.stats-value {
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: var(--tp-text, #212529);
-}
-
-.connections-list {
-  max-height: 200px;
-  overflow-y: auto;
-  background: var(--tp-bg-subtle, #f8f9fa);
-  border-radius: 0.25rem;
-  padding: 0.5rem;
-}
-
-.connection-item {
-  padding: 0.25rem 0;
-  font-size: 0.875rem;
-  font-family: monospace;
-}
-
-.protocol-breakdown {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.protocol-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.5rem;
-  background: var(--tp-bg-subtle, #f8f9fa);
-  border-radius: 0.25rem;
-  font-size: 0.875rem;
-  color: var(--tp-text, #212529);
-}
-
-.protocol-name {
-  font-weight: 500;
-}
-
-.text-monospace {
-  font-family: monospace;
-}
-
-.node-details-hostname {
-  font-size: 0.85rem;
-}
-
-.node-details-evidence {
-  font-size: 0.75rem;
-}
-
-.node-details-peer-row {
-  cursor: pointer;
-}
-
-.node-details-peer-icon {
-  font-size: 0.7rem;
-}
-
-/* Nested modals (add-evidence, evidence explainer) shown from within the NodeDetails panel, which
-   itself renders as a raised overlay in monitor mode (z-index ~1070). These must sit above it so
-   they aren't hidden behind the node panel (modal-in-modal). React-Bootstrap sets its own INLINE
-   z-index on the .modal/.modal-backdrop elements, so !important is required to win over it. */
-.tp-nested-modal {
-  z-index: 2000 !important;
-}
-.tp-nested-modal-backdrop {
-  z-index: 1990 !important;
-}
