--- frontend/src/components/network/NodeDetails/index.ts
@@ -1,1 +1,0 @@
-export { NodeDetails } from './NodeDetails';
